Refactor redux file structure to make it more scalable. - Create separate file for actions. - Create 'actionTypes.js' and 'books.js' under actions directory. - Update paths and imports where applicable.
