+++ src/redux/actions/actionTypes.js
@@ -1,0 +1,2 @@
+export const ADD_BOOK = 'bookStore/books/ADD_BOOK';
+export const REMOVE_BOOK = 'bookStore/books/REMOVE_BOOK';
