+++ src/redux/actions/books.js
@@ -1,0 +1,11 @@
+import { ADD_BOOK, REMOVE_BOOK } from './actionTypes';
+
+export const addBook = (payload) => ({
+  type: ADD_BOOK,
+  payload,
+});
+
+export const removeBook = (payload) => ({
+  type: REMOVE_BOOK,
+  payload,
+});
