+++ src/redux/reducers/books/books.js
@@ -1,20 +1,8 @@
 /* eslint-disable consistent-return */
-import getBooks from '../../api/api';
-
-const ADD_BOOK = 'bookStore/books/ADD_BOOK';
-const REMOVE_BOOK = 'bookStore/books/REMOVE_BOOK';
+import { ADD_BOOK, REMOVE_BOOK } from '../../actions/actionTypes';
+import getBooks from '../../../api/api';
 
 const initialState = getBooks();
-
-export const addBook = (payload) => ({
-  type: ADD_BOOK,
-  payload,
-});
-
-export const removeBook = (payload) => ({
-  type: REMOVE_BOOK,
-  payload,
-});
 
 const reducer = (state = initialState, action) => {
   switch (action.type) {
